Refactror: Rename original "error.rs" to "py_error.rs"
