+++ crates/circuit/src/lib.rs
@@ -23,7 +23,6 @@ pub mod dag_circuit;
 pub mod dag_node;
 mod dot_utils;
 pub mod duration;
-pub mod error;
 pub mod gate_matrix;
 pub mod imports;
 pub mod interner;
@@ -33,6 +32,7 @@ pub mod operations;
 pub mod packed_instruction;
 pub mod parameter;
 pub mod parameter_table;
+pub mod py_error;
 pub mod register_data;
 pub mod slice;
 pub mod util;
