+++ crates/circuit/src/py_error.rs
